TST: isolate environment unit tests
